requirements.txt
numpy==1.23.4
Pillow==9.3.0
PySimpleGUI==4.60.4
scipy==1.9.3
